@@ -2,7 +2,7 @@
   "name": "image-bot",
   "version": "1.0.0",
   "description": "A Discord bot which posts images from user searches.",
-  "main": "image-bot.js",
+  "main": "image-bot.ts",
   "author": "steelcolossus",
   "repository": {
     "type": "git",
@@ -10,18 +10,27 @@
   },
   "private": true,
   "scripts": {
-    "start": "node .",
+    "build:prod": "tsc --project tsconfig.prod.json",
+    "start": "ts-node .",
+    "start:prod": "node ./built/image-bot.js",
     "test": "echo Error: this module has no tests && exit 1",
-    "lint": "eslint **/*.js",
-    "lint:fix": "eslint **/*.js --fix"
+    "lint": "eslint **/*.ts",
+    "lint:fix": "eslint **/*.ts --fix"
   },
   "dependencies": {
-    "discord.js": "^11.6.4",
+    "discord.js": "^12.5.1",
     "dotenv": "^8.2.0",
     "google-images": "^2.1.0",
     "winston": "^3.3.3"
   },
   "devDependencies": {
-    "eslint": "^7.17.0"
+    "@types/google-images": "^2.1.0",
+    "@types/node": "^14.14.21",
+    "@types/ws": "^7.4.0",
+    "@typescript-eslint/eslint-plugin": "^4.13.0",
+    "@typescript-eslint/parser": "^4.13.0",
+    "eslint": "^7.18.0",
+    "ts-node": "^9.1.1",
+    "typescript": "^4.1.3"
   }
 }
